+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/annotation-generator
 
+## 0.1.10
+
+### Patch Changes
+
+-   Updated dependencies [173b5f2]
+    -   @sap-ux/project-access@1.25.4
+    -   @sap-ux/fiori-annotation-api@0.1.20
+
+## 0.1.9
+
+### Patch Changes
+
+-   Updated dependencies [e7b9184]
+    -   @sap-ux/project-access@1.25.3
+    -   @sap-ux/fiori-annotation-api@0.1.19
+
+## 0.1.8
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.25.2
+-   @sap-ux/fiori-annotation-api@0.1.18
+
 ## 0.1.7
 
 ### Patch Changes